refactor: Extract TransactionModalRow
