+++ src/ducks/transactions/TransactionModal.styl
@@ -3,16 +3,6 @@
 .Separated > * + *
     border-top 1px solid var(--silver)
 
-.TransactionModalRow
-    min-height 57px
-    box-sizing border-box
-    padding: 12px 24px
-
-.TransactionModalRow-disabled
-    color var(--coolGrey)
-    font-style italic
-    cursor default
-
 .TransactionModalRowIcon-reversed
     transform rotate(180deg)
