+++ src/ducks/transactions/TransactionModalRow.styl
@@ -1,0 +1,10 @@
+
+.TransactionModalRow
+    min-height 57px
+    box-sizing border-box
+    padding: 12px 24px
+
+.TransactionModalRow-disabled
+    color var(--coolGrey)
+    font-style italic
+    cursor default
