+++ src/ducks/transactions/actions/KonnectorAction/index.jsx
@@ -6,7 +6,7 @@ import ButtonAction from 'cozy-ui/react/ButtonAction'
 import flag from 'cozy-flags'
 import icon from 'assets/icons/actions/icon-link-out.svg'
 import styles from 'ducks/transactions/TransactionActions.styl'
-import { TransactionModalRow } from 'ducks/transactions/TransactionModal'
+import TransactionModalRow from 'ducks/transactions/TransactionModalRow'
 import palette from 'cozy-ui/react/palette'
 import { triggersConn } from 'doctypes'
 import InformativeModal from 'ducks/transactions/actions/KonnectorAction/InformativeModal'
