+++ src/ducks/transactions/actions/UrlLinkAction.jsx
@@ -5,7 +5,7 @@ import Chip from 'cozy-ui/react/Chip'
 import Icon from 'cozy-ui/react/Icon'
 import flag from 'cozy-flags'
 import styles from 'ducks/transactions/TransactionActions.styl'
-import { TransactionModalRow } from 'ducks/transactions/TransactionModal'
+import TransactionModalRow from 'ducks/transactions/TransactionModalRow'
 import palette from 'cozy-ui/react/palette'
 
 const name = 'url'
